Windows-Client/requirements.txt
requests>=2.31.0
Pillow>=10.2.0
pyautogui>=0.9.54
keyboard>=0.13.5
psutil>=5.9.5
pywin32>=306
python-dotenv>=1.0.0
schedule>=1.2.0 